% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alimentation}
\alias{alimentation}
\title{Food consumption in European countries}
\format{
An object of class \code{data.frame} with 25 rows and 13 columns.
}
\usage{
alimentation
}
\description{
The `alimentation` data set contains the percentage composition of food
consumption in 25 European countries during the 1980s. The food categories
are:

\itemize{
  \item `RM`: red meat (pork, veal, beef),
  \item `WM`: white meat (chicken),
  \item `E`: eggs,
  \item `M`: milk,
  \item `F`: fish,
  \item `C`: cereals,
  \item `S`: starch (potatoes),
  \item `N`: nuts,
  \item `FV`: fruits and vegetables.
}

The data set also contains categorical variables indicating whether the
country belongs to the North or South/Mediterranean group, and whether it is
an Eastern or Western European country.
}
\keyword{datasets}
