% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{alr_basis}
\alias{alr_basis}
\title{Additive log-ratio basis}
\usage{
alr_basis(dim, denominator = NULL, numerator = NULL)
}
\arguments{
\item{dim}{Number of parts. It can be a single integer, a matrix or data
frame, or a character vector of part names.}

\item{denominator}{Part used as denominator. By default, the last part is
used.}

\item{numerator}{Parts used as numerators. By default, all parts except the
denominator are used, preserving their original order.}
}
\value{
A matrix defining the alr coordinate system.
}
\description{
Construct the transformation matrix associated with additive log-ratio (alr)
coordinates.
}
\examples{
alr_basis(5)
alr_basis(5, 3)
alr_basis(5, 3, c(1, 5, 2, 4))

}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
Chapman & Hall, London.
}
