% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arctic_lake}
\alias{arctic_lake}
\title{Arctic lake sediments at different depths}
\format{
An object of class \code{data.frame} with 39 rows and 5 columns.
}
\usage{
arctic_lake
}
\description{
The `arctic_lake` data set records the three-part composition
\eqn{[sand, silt, clay]} of 39 sediment samples collected at different water
depths in an Arctic lake.
}
\keyword{datasets}
