% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{basis}
\alias{basis}
\title{Coordinates basis}
\usage{
basis(H)
}
\arguments{
\item{H}{coordinates for which basis should be shown}
}
\value{
basis used to create coordinates H
}
\description{
Obtain coordinates basis
}
