% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blood_mn}
\alias{blood_mn}
\title{The MN blood system}
\format{
An object of class \code{data.frame} with 49 rows and 5 columns.
}
\usage{
blood_mn
}
\description{
In humans, the main blood group systems are the ABO system, the Rh system,
and the MN system. The MN blood system is related to proteins of the red
blood cell plasma membrane. Its inheritance pattern is autosomal with
codominance, meaning that the heterozygous phenotype is distinct from both
homozygous phenotypes.

The three phenotypes are M, N, and MN. Their frequencies vary across
populations. Under the Hardy-Weinberg principle, allele and genotype
frequencies remain constant across generations in the absence of evolutionary
forces, implying that

\eqn{\frac{x_{MM} x_{NN}}{x_{MN}^2} = \frac{1}{4}}{
(x_MM x_NN) / x_MN^2 = 1/4
}

where \eqn{x_{MM}} and \eqn{x_{NN}} are the genotype frequencies of the
homozygotes and \eqn{x_{MN}} is the genotype frequency of heterozygotes.
}
\keyword{datasets}
