% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi_activity}
\alias{bmi_activity}
\title{Physical activity and body mass index}
\format{
An object of class \code{data.frame} with 393 rows and 8 columns.
}
\usage{
bmi_activity
}
\description{
The `bmi_activity` data set records the proportion of daily time spent in
sleep (`sleep`), sedentary behaviour (`sedent`), light physical activity
(`Lpa`), moderate physical activity (`Mpa`), and vigorous physical activity
(`Vpa`) for 393 children. The standardized body mass index (`zBMI`) of each
child is also included.

This data set was used in the example of Dumuid et al. (2019) to examine the
expected differences in zBMI associated with reallocations of daily time
between sleep, sedentary behaviour, and physical activity. Because the
original data are confidential, `bmi_activity` contains simulated data that
mimic the main features of the original study.
}
\references{
Dumuid, D., Pedisic, Z., Stanford, T. E., Martín-Fernández, J. A., Hron, K.,
Maher, C., Lewis, L. K., & Olds, T. S. (2019).
\emph{The Compositional Isotemporal Substitution Model: a Method for
Estimating Changes in a Health Outcome for Reallocation of Time between
Sleep, Sedentary Behaviour, and Physical Activity}.
Statistical Methods in Medical Research, \strong{28}(3), 846--857.
}
\keyword{datasets}
