% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{cbalance_approx}
\alias{cbalance_approx}
\title{Balance generated from the first canonical correlation component}
\usage{
cbalance_approx(Y, X)
}
\arguments{
\item{Y}{compositional dataset}

\item{X}{explanatory dataset}
}
\value{
matrix
}
\description{
Balance generated from the first canonical correlation component
}
