% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{cc_basis}
\alias{cc_basis}
\title{Canonical-correlation log-ratio basis}
\usage{
cc_basis(Y, X)
}
\arguments{
\item{Y}{A compositional data set.}

\item{X}{An explanatory data set.}
}
\value{
A matrix whose columns define a canonical-correlation-oriented ilr
  basis.
}
\description{
Construct an ilr basis rotated according to canonical correlations between a
compositional response data set and an explanatory data set.
}
