% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{cdp_basis}
\alias{cdp_basis}
\title{CoDaPack default ilr basis}
\usage{
cdp_basis(dim)
}
\arguments{
\item{dim}{Number of parts. It can be a single integer, a matrix or data
frame, or a character vector of part names.}
}
\value{
A matrix with \eqn{D} rows and \eqn{D - 1} columns containing the
  CoDaPack default ilr basis.
}
\description{
Construct the default isometric log-ratio basis used in CoDaPack.
}
\examples{
cdp_basis(5)
cdp_basis(c("a", "b", "c", "d"))

}
