% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cdp_partition}
\alias{cdp_partition}
\title{CoDaPack's default binary partition}
\usage{
cdp_partition(ncomp)
}
\arguments{
\item{ncomp}{number of parts}
}
\value{
matrix
}
\description{
Compute the default binary partition used in CoDaPack's software
}
\examples{
cdp_partition(4)
}
