% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{clr_basis}
\alias{clr_basis}
\title{Centered log-ratio basis}
\usage{
clr_basis(dim)
}
\arguments{
\item{dim}{Number of parts. It can be a single integer, a matrix or data
frame, or a character vector of part names.}
}
\value{
A square matrix defining the clr coordinate system.
}
\description{
Construct the transformation matrix associated with centered log-ratio (clr)
coordinates.
}
\details{
CLR coordinates are linearly dependent and lie in the \eqn{D - 1}
dimensional clr-plane.
}
\examples{
B <- clr_basis(5)
clr_coordinates <- coordinates(c(1, 2, 3, 4, 5), B)
sum(clr_coordinates) < 1e-15

}
\references{
Aitchison, J. (1986). \emph{The Statistical Analysis of Compositional Data}.
Chapman & Hall, London.
}
