% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.base.R
\docType{package}
\name{coda.base}
\alias{coda.base}
\title{coda.base}
\description{
A minimum set of functions to perform compositional data analysis
using the log-ratio approach introduced by John Aitchison (1982)
<https://www.jstor.org/stable/2345821>. Main functions
have been implemented in c++ for better performance.
}
\author{
Marc Comas-Cufí
}
