% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{composition}
\alias{composition}
\alias{comp}
\title{Compositions from coordinates with respect to a basis}
\usage{
composition(H, basis = "ilr")

comp(H, basis = "ilr")
}
\arguments{
\item{H}{Coordinates of a composition. It can be a numeric matrix, a data
frame, or a numeric vector.}

\item{basis}{Basis used to interpret the coordinates. Either a character
string naming a predefined basis or a matrix.}
}
\value{
A composition corresponding to the given coordinates.
}
\description{
Reconstruct a composition from coordinates with respect to a given basis.
}
\seealso{
\code{\link{coordinates}}, \code{\link{ilr_basis}}, \code{\link{alr_basis}},
\code{\link{clr_basis}}, \code{\link{sbp_basis}}
}
