% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_basis_constrained.R
\name{constrained_pb}
\alias{constrained_pb}
\title{Constrained principal balance basis}
\usage{
constrained_pb(X, angle = FALSE)
}
\arguments{
\item{X}{Compositional data set.}

\item{angle}{Logical; if `TRUE`, use the angle criterion instead of the
variance criterion.}
}
\value{
A matrix whose columns are constrained principal balances.
}
\description{
Compute a basis of constrained principal balances recursively.
}
\keyword{internal}
