% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\alias{coord}
\alias{alr_c}
\alias{clr_c}
\alias{ilr_c}
\alias{olr_c}
\title{Coordinates of compositions with respect to a basis}
\usage{
coordinates(X, basis = "ilr")

coord(..., basis = "ilr")

alr_c(X)

clr_c(X)

ilr_c(X)

olr_c(X)
}
\arguments{
\item{X}{A compositional data set. It can be a numeric matrix, a data frame,
or a numeric vector.}

\item{basis}{Basis used to compute the coordinates. Either a character string
naming a predefined basis or a matrix with log-ratio basis vectors in
columns.}

\item{...}{components of the composition}
}
\value{
Coordinates of `X` with respect to the given `basis`. The returned object has
the same general type as the input when possible.
}
\description{
Compute coordinates of a composition or a compositional data set with respect
to a given log-ratio basis.

The `basis` argument can be either:
\itemize{
  \item a character string identifying a predefined coordinate system, or
  \item a matrix whose columns define a system of log-contrasts.
}

The predefined options are:
\itemize{
  \item `"ilr"`: isometric log-ratio coordinates,
  \item `"olr"`: orthonormal log-ratio coordinates,
  \item `"clr"`: centered log-ratio coordinates,
  \item `"alr"`: additive log-ratio coordinates,
  \item `"pw"`: pairwise log-ratios,
  \item `"pc"`: principal component log-ratio coordinates,
  \item `"pb"`: principal balance coordinates,
  \item `"cdp"`: CoDaPack default balances.
}
}
\examples{
coordinates(1:5)

B <- ilr_basis(5)
coordinates(1:5, B)

X <- rbind(1:5, 2:6)
coordinates(X, "clr")

}
\seealso{
\code{\link{ilr_basis}}, \code{\link{alr_basis}}, \code{\link{clr_basis}},
\code{\link{sbp_basis}}, \code{\link{composition}}
}
