% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eurostat_employment}
\alias{eurostat_employment}
\title{Employment distribution in EUROSTAT countries}
\format{
An object of class \code{data.frame} with 29 rows and 17 columns.
}
\usage{
eurostat_employment
}
\description{
According to the three-sector theory, employment shifts from the primary
sector (raw material extraction), to the secondary sector (industry, energy,
and construction), and then to the tertiary sector (services) as economies
develop. The `eurostat_employment` data set contains EUROSTAT data on
employment, aggregated for both sexes and all ages, distributed by economic
activity in 2008 for 29 EUROSTAT member countries.

A related variable is the logarithm of gross domestic product per person in
EUR at current prices (`logGDP`). For exploratory purposes, it is also
categorised as a binary variable indicating values above or below the median
(`Binary GDP`).

The employment composition has 11 parts:
\itemize{
  \item Primary sector
  \item Manufacturing
  \item Energy
  \item Construction
  \item Trade repair transport
  \item Hotels restaurants
  \item Financial intermediation
  \item Real estate
  \item Educ admin defense soc sec
  \item Health social work
  \item Other services
}
}
\keyword{datasets}
