% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foraminiferals}
\alias{foraminiferals}
\title{Paleocological compositions}
\format{
An object of class \code{data.frame} with 30 rows and 5 columns.
}
\usage{
foraminiferals
}
\description{
The `foraminiferals` data set (Aitchison, 1986) is a classical example of
paleocological compositional data. It contains the composition of four fossil
types (Neogloboquadrina atlantica, Neogloboquadrina pachyderma,
Globorotalia obesa, and Globigerinoides triloba) at 30 different depths.

Because the data contain rounded zeros, zero-replacement techniques are
typically required before analysis. A natural goal is then to study the
association between fossil composition and depth.
}
\keyword{datasets}
