% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gmean}
\alias{gmean}
\title{Geometric Mean}
\usage{
gmean(x, zero.rm = FALSE, trim = 0, na.rm = FALSE)
}
\arguments{
\item{x}{A nonnegative vector.}

\item{zero.rm}{a logical value indicating whether zero values should be stripped
before the computation proceeds.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each
end of x before the mean is computed. Values of trim outside that range are
taken as the nearest endpoint.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\description{
Generic function for the (trimmed) geometric mean.
}
\seealso{
\code{\link{center}}
}
