% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_expend}
\alias{house_expend}
\title{Household expenditures}
\format{
An object of class \code{data.frame} with 27 rows and 15 columns.
}
\usage{
house_expend
}
\description{
The `house_expend` data set, obtained from Eurostat, records the composition
of mean household consumption expenditure across 12 expenditure categories in
27 European Union countries. Some values are rounded zeros.

In addition, the data set contains gross domestic product values for 2005
(`GDP05`) and 2014 (`GDP14`). A relevant analysis is the relationship
between expenditure compositions and GDP.
}
\keyword{datasets}
