% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_budget}
\alias{household_budget}
\title{Household budget patterns}
\format{
An object of class \code{data.frame} with 40 rows and 6 columns.
}
\usage{
household_budget
}
\description{
In a sample survey of single persons living alone in rented accommodation,
twenty men and twenty women were randomly selected and asked to record their
expenditure over one month in the following four mutually exclusive and
exhaustive commodity groups:

\itemize{
  \item `Hous`: housing, including fuel and light,
  \item `Food`: foodstuffs, including alcohol and tobacco,
  \item `Serv`: services, including transport and vehicles,
  \item `Other`: other goods, including clothing, footwear, and durable goods.
}
}
\keyword{datasets}
