% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kilauea_iki}
\alias{kilauea_iki}
\title{Chemical composition of volcanic rocks from Kilauea Iki}
\format{
A data frame with 17 observations and 11 variables:
\describe{
  \item{SiO2}{Silicon dioxide}
  \item{TiO2}{Titanium dioxide}
  \item{Al2O3}{Aluminium oxide}
  \item{Fe2O3}{Ferric oxide}
  \item{FeO}{Ferrous oxide}
  \item{MnO}{Manganese oxide}
  \item{MgO}{Magnesium oxide}
  \item{CaO}{Calcium oxide}
  \item{Na2O}{Sodium oxide}
  \item{K2O}{Potassium oxide}
  \item{P2O5}{Phosphorus pentoxide}
}
}
\source{
Richter, D. H., & Moore, J. G. (1966). \emph{Petrology of Kilauea Iki lava
lake, Hawaii}. Geological Survey Professional Paper 537-B.
}
\usage{
kilauea_iki
}
\description{
The `kilauea_iki` data set contains the chemical composition of volcanic
rocks sampled from the lava lake at Kilauea Iki (Hawaii). The data represent
major oxide concentrations in fractional form.
}
\details{
The variability in oxide concentrations is attributed to magnesian olivine
fractionation from a single magmatic mass, as suggested by Richter and Moore
(1966).
}
\keyword{datasets}
