% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals_milk}
\alias{mammals_milk}
\title{Mammals' milk}
\format{
An object of class \code{data.frame} with 24 rows and 6 columns.
}
\usage{
mammals_milk
}
\description{
The `mammals_milk` data set contains the percentages of five constituents of
the milk of 24 mammals:
\eqn{[W, P, F, L, A]}{[W, P, F, L, A]},
where `W` is water, `P` is protein, `F` is fat, `L` is lactose, and `A` is
ash.
}
\keyword{datasets}
