% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk_cows}
\alias{milk_cows}
\title{Milk composition study}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 116 rows and 10 columns.
}
\usage{
milk_cows
}
\description{
In an attempt to improve the quality of cow milk, milk from thirty cows was
assessed before and after a controlled dietary and hormonal regime over eight
weeks. A control group of thirty cows kept under the usual regime was also
included.

The `milk_cows` data set provides the complete before/after milk composition
data for the sixty cows, with the proportions of protein (`pr`), milk fat
(`mf`), carbohydrate (`ch`), calcium (`Ca`), sodium (`Na`), and potassium
(`K`).
}
\keyword{datasets}
