% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{montana}
\alias{montana}
\title{Concentration of minor elements in coal ashes}
\format{
An object of class \code{data.frame} with 229 rows and 6 columns.
}
\usage{
montana
}
\description{
The `montana` data set contains 229 samples of the concentration (in ppm) of
five minor elements \eqn{[Cr, Cu, Hg, U, V]} in coal ashes from the Fort
Union formation (Montana, USA), in the Powder River Basin.

The five measured elements form a fully observed subcomposition of a much
larger chemical composition. Since the data are given in parts per million and
all concentrations were measured, a residual component could in principle be
added to close the compositions to \eqn{10^6}.
}
\keyword{datasets}
