% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{pairwise_basis}
\alias{pairwise_basis}
\title{Pairwise log-ratio generating system}
\usage{
pairwise_basis(dim)
}
\arguments{
\item{dim}{Number of parts. It can be a single integer, a matrix or data
frame, or a character vector of part names.}
}
\value{
A matrix, or a sparse matrix for large dimensions, whose columns
  represent all pairwise log-ratio generators.
}
\description{
Construct the system of all pairwise log-ratios between parts.
}
