% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parliament2017}
\alias{parliament2017}
\title{Catalan Parliament election results in 2017 by region}
\format{
A data frame with 42 rows and 9 variables:
\describe{
  \item{com}{Region}
  \item{cs}{Votes for the Ciutadans party}
  \item{jxcat}{Votes for the Junts per Catalunya party}
  \item{erc}{Votes for the Esquerra Republicana de Catalunya party}
  \item{psc}{Votes for the Partit Socialista de Catalunya party}
  \item{catsp}{Votes for the Catalunya Sí que es Pot party}
  \item{cup}{Votes for the Candidatura d'Unitat Popular party}
  \item{pp}{Votes for the Partit Popular party}
  \item{other}{Votes for other parties}
}
}
\source{
Idescat, statistics on Catalan Parliament elections.
}
\usage{
parliament2017
}
\description{
The `parliament2017` data set contains the results of the 2017 Catalan
Parliament election aggregated by region.
}
\keyword{datasets}
