% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{pc_basis}
\alias{pc_basis}
\title{Principal component log-ratio basis}
\usage{
pc_basis(X)
}
\arguments{
\item{X}{Compositional data set.}
}
\value{
A matrix whose columns define a principal-component-oriented ilr
  basis.
}
\description{
Construct an ilr basis rotated according to the principal components of the
log-ratio coordinates of a compositional data set.
}
