% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{petrafm}
\alias{petrafm}
\title{Calc-alkaline and tholeiitic volcanic rocks}
\format{
An object of class \code{data.frame} with 100 rows and 4 columns.
}
\usage{
petrafm
}
\description{
The `petrafm` data set contains 100 classified volcanic rock samples from
Ontario (Canada). The three-part composition is

\eqn{[A: Na_2O + K_2O;\ F: FeO + 0.8998\,Fe_2O_3;\ M: MgO]}{
[A: Na2O + K2O; F: FeO + 0.8998 Fe2O3; M: MgO]
}

Rocks from the calc-alkaline magma series (25 samples) can be distinguished
from those of the tholeiitic magma series (75 samples) using an AFM diagram.
}
\keyword{datasets}
