% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollen}
\alias{pollen}
\title{Pollen composition in fossils}
\format{
An object of class \code{data.frame} with 30 rows and 4 columns.
}
\usage{
pollen
}
\description{
The `pollen` data set contains 30 fossil pollen samples from three different
locations (recorded in variable `group`). The measured composition is the
three-part composition \eqn{[pinus, abies, quercus]}.
}
\keyword{datasets}
