% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{print.coda}
\alias{print.coda}
\title{Printing coordinates}
\usage{
\method{print}{coda}(x, ..., basis = getOption("coda.base.basis"))
}
\arguments{
\item{x}{coordinates}

\item{...}{parameters passed to print function}

\item{basis}{boolean to show or not the basis with the output}
}
\description{
The function hides the basis attribute. An option is included to
show such basis.
}
