% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{serprot}
\alias{serprot}
\title{Serum proteins}
\format{
An object of class \code{data.frame} with 36 rows and 7 columns.
}
\usage{
serprot
}
\description{
The `serprot` data set records the percentages of four serum proteins from
blood samples of 30 patients. Fourteen patients have one disease and sixteen
have another.

The four-part compositions are formed by
\eqn{[albumin, pre\text{-}albumin, globulin\ A, globulin\ B]}{
[albumin, pre-albumin, globulin A, globulin B]
}.
}
\keyword{datasets}
