% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statistician_time}
\alias{statistician_time}
\title{A statistician's time budget}
\format{
An object of class \code{data.frame} with 20 rows and 7 columns.
}
\usage{
statistician_time
}
\description{
The `statistician_time` data set records the daily time budget of an academic
statistician across 20 working days. The six activities are teaching (`T`),
consultation (`C`), administration (`A`), research (`R`), other wakeful
activities (`O`), and sleep (`S`).

These activities may also be grouped into work (`T`, `C`, `A`, `R`) and
leisure (`O`, `S`). The data allow investigation of the relationship between
detailed time-allocation patterns and the broader division between work and
leisure.
}
\keyword{datasets}
