% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variation_array}
\alias{variation_array}
\title{Variation array is returned.}
\usage{
variation_array(X, include_means = FALSE, ml_covariance = FALSE)
}
\arguments{
\item{X}{Compositional dataset}

\item{include_means}{if TRUE logratio means are included in the lower-left triangle}

\item{ml_covariance}{if TRUE Maximum-likelihood estimation of the covariance for the multivariate normal distribution is used (dividing the scatter matrix by n instead of n-1)}
}
\value{
variation array matrix
}
\description{
Variation array is returned.
}
\examples{
set.seed(1)
X = matrix(exp(rnorm(5*100)), nrow=100, ncol=5)
variation_array(X)
variation_array(X, include_means = TRUE)
}
