% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weibo_hotels}
\alias{weibo_hotels}
\title{Hotel posts in social media}
\format{
An object of class \code{data.frame} with 10 rows and 5 columns.
}
\usage{
weibo_hotels
}
\description{
The `weibo_hotels` data set compares the use of Weibo (the Chinese equivalent
of Facebook) in hospitality e-marketing between small and medium
establishments and larger hotel businesses in China.

The 50 latest posts from the Weibo page of each hotel (\eqn{n = 10}) were
content-analysed and coded into a four-part composition:
\eqn{[facilities, food, events, promotions]}.
Hotels were also classified by size as large (`L`) or small (`S`).
}
\keyword{datasets}
