% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alimentation}
\alias{alimentation}
\title{Food consumption in European countries}
\format{
An object of class \code{data.frame} with 25 rows and 13 columns.
}
\usage{
alimentation
}
\description{
The alimentation data set contains the percentages of consumption of several
types of food in 25 European countries during the 80s. The categories are:
* RM: red meat (pork, veal, beef),
* WM: white meat (chicken),
* E: eggs,
* M: milk,
* F: fish,
* C: cereals,
* S: starch (potatoes),
* N: nuts, and
* FV: fruits and vegetables.
}
\details{
Moreover, the dataset contains a categorical variable that
shows if the country is from the North or a Southern Mediterranean country. In
addition, the countries are classified as Eastern European or as Western European.
}
\keyword{datasets}
