% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{alr_basis}
\alias{alr_basis}
\title{Build an additive log-ratio basis}
\usage{
alr_basis(dim, denominator = dim, numerator = which(denominator != 1:dim))
}
\arguments{
\item{dim}{number of components}

\item{denominator}{part used as denominator}

\item{numerator}{parts to be used as numerator. By default all except the denominator following same order.}
}
\value{
matrix
}
\description{
Build an additive log-ratio basis
}
\examples{
alr_basis(5)
}
