% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi_activity}
\alias{bmi_activity}
\title{Physical activity and body mass index}
\format{
An object of class \code{data.frame} with 393 rows and 8 columns.
}
\usage{
bmi_activity
}
\description{
Physical activity and body mass index
}
\keyword{datasets}
