% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{cc_basis}
\alias{cc_basis}
\title{Isometric log-ratio basis based on canonical correlations}
\usage{
cc_basis(Y, X)
}
\arguments{
\item{Y}{compositional dataset}

\item{X}{explanatory dataset}
}
\value{
matrix
}
\description{
Isometric log-ratio basis based on canonical correlations
}
