% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{cc_basis}
\alias{cc_basis}
\title{Isometric Log-Ratio Basis Based on Canonical Correlations}
\usage{
cc_basis(Y, X)
}
\arguments{
\item{Y}{A compositional dataset (matrix or data frame).}

\item{X}{An explanatory dataset (matrix or data frame).}
}
\value{
A matrix representing the isometric log-ratio basis.
}
\description{
Constructs an isometric log-ratio (ilr) basis for a compositional dataset,
optimized with respect to canonical correlations with an explanatory dataset.
}
