% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{cdp_basis}
\alias{cdp_basis}
\title{Isometric log-ratio basis based on Balances.}
\usage{
cdp_basis(dim)
}
\arguments{
\item{dim}{dimension to build the ILR basis based on balanced balances}
}
\value{
matrix
}
\description{
The function return default balances used in CoDaPack software.
}
