% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{clr_basis}
\alias{clr_basis}
\title{Build a centered log-ratio basis}
\usage{
clr_basis(dim)
}
\arguments{
\item{dim}{number of components}
}
\value{
matrix
}
\description{
Build a centered log-ratio basis
}
\examples{
clr_basis(5)
}
