% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_replacement_imputation.R
\name{coda_replacement}
\alias{coda_replacement}
\title{Replacement of missing values and below-detection zeros in compositional data}
\usage{
coda_replacement(
  X,
  DL = NULL,
  dl_prop = 0.65,
  eps = 1e-04,
  parameters = FALSE,
  debug = FALSE,
  maxit = 500
)
}
\arguments{
\item{X}{A compositional data set: numeric matrix or data frame where rows
represent observations and columns represent parts.}

\item{DL}{An optional matrix or vector of detection limits. If `NULL`, the
minimum non-zero value in each column of `X` is used.}

\item{dl_prop}{A numeric value between 0 and 1 used for initialization in
the EM algorithm.}

\item{eps}{Convergence tolerance.}

\item{parameters}{Logical; if `TRUE`, return additional estimated parameters.}

\item{debug}{Logical; if `TRUE`, print the log-likelihood at each iteration.}

\item{maxit}{Maximum number of iterations}
}
\value{
If `parameters = FALSE`, the imputed object with the same format as `X`
(`matrix` or `data.frame`, preserving data-frame subclasses when possible)
and preserving original names.
If `parameters = TRUE`, a list with the estimated clr mean, clr covariance,
and imputed clr coordinates.
}
\description{
Performs imputation of missing values and/or values below the detection limit
in compositional data using an EM algorithm assuming normality on the simplex.
}
\examples{
X <- matrix(c(
  0.00, 0.30, 0.70,
  0.20,   NA, 0.80,
  0.40, 0.60, 0.00,
  0.25, 0.25, 0.50,
  0.10, 0.30, 0.60
), ncol = 3, byrow = TRUE)
colnames(X) <- c("sand", "silt", "clay")

DL <- c(0.05, 0.05, 0.05)

X_imp <- coda_replacement(X, DL = DL, maxit = 20)
X_imp

set.seed(1)
X <- composition(matrix(rnorm(5*50), ncol = 5))
X[sample(c(T,F), 6*50, replace = TRUE, c(0.3, 0.7))] <- NA
params <- coda_replacement(X, parameters = TRUE, debug = TRUE)
names(params)
params$clr_mu
composition(params$clr_h)
}
