% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_replacement_imputation.R
\name{coda_replacement}
\alias{coda_replacement}
\title{Replacement of missing values and below-detection zeros in compositional data}
\usage{
coda_replacement(
  X,
  DL = NULL,
  dl_prop = 0.65,
  eps = 1e-04,
  parameters = FALSE,
  debug = FALSE,
  maxit = 500
)
}
\arguments{
\item{X}{A compositional data set: numeric matrix or data frame where rows
represent observations and columns represent parts.}

\item{DL}{An optional matrix or vector of detection limits. If `NULL`, the
minimum non-zero value in each column of `X` is used.}

\item{dl_prop}{A numeric value between 0 and 1 used for initialization in
the EM algorithm.}

\item{eps}{Convergence tolerance.}

\item{parameters}{Logical; if `TRUE`, return additional estimated parameters.}

\item{debug}{Logical; if `TRUE`, print the log-likelihood at each iteration.}

\item{maxit}{Maximum number of iterations}
}
\value{
If `parameters = FALSE`, a numeric matrix with imputed values.
If `parameters = TRUE`, a list with the estimated clr mean, clr covariance,
and imputed clr coordinates.
}
\description{
Performs imputation of missing values and/or values below the detection limit
in compositional data using an EM algorithm assuming normality on the simplex.
}
