% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{composition}
\alias{composition}
\alias{comp}
\title{Get composition from coordinates w.r.t.  an specific basis}
\usage{
composition(H, basis = "ilr")

comp(H, basis = "ilr")
}
\arguments{
\item{H}{coordinates of a composition. Either a matrix, a data.frame or a vector}

\item{basis}{basis used to calculate the coordinates}
}
\value{
coordinates with respect the given basis
}
\description{
Calculate a composition from coordinates with respect a given basis
}
\seealso{
See functions \code{\link{ilr_basis}}, \code{\link{alr_basis}},
\code{\link{clr_basis}}, \code{\link{sbp_basis}}
to define different compositional basis.
See function \code{\link{coordinates}} to obtain details on how to calculate
coordinates of a given composition.
}
