% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\alias{coord}
\title{Get coordinates from compositions w.r.t. an specific basis}
\usage{
coordinates(X, basis = "ilr", basis_return = TRUE)

coord(..., basis = "ilr")
}
\arguments{
\item{X}{compositional dataset. Either a matrix, a data.frame or a vector}

\item{basis}{basis used to calculate the coordinates. \code{basis} can be either a string or a matrix.
Accepted values for strings are: 'ilr' (default), 'clr', 'alr', 'pw', 'pc', 'pb' and 'cdp'. If \code{basis} is a matrix, it is expected
to have log-ratio basis given in columns.}

\item{basis_return}{Should the basis be returned as attribute? (default: \code{TRUE})}

\item{...}{components of the compositional data}
}
\value{
Coordinates of composition \code{X} with respect the given \code{basis}.
}
\description{
Calculate the coordinates of a composition with respect a given basis
}
\details{
\code{coordinates} function calculates the coordinates of a compositiona w.r.t. a given basis. `basis` parameter is
used to set the basis, it can be either a matrix defining the log-contrasts in columns or a string defining some well-known
log-contrast: 'alr' 'clr', 'ilr', 'pw', 'pc', 'pb' and 'cdp', for the additive log-ratio, centered log-ratio, isometric log-ratio,
pairwise log-ratio, clr principal components, clr principal balances or default's CoDaPack balances respectively.
}
\examples{
coordinates(c(1,2,3,4,5))
h = coordinates(c(1,2,3,4,5))
basis(h)
# basis is shown if 'coda.base.basis' option is set to TRUE
options('coda.base.basis' = TRUE)
coordinates(c(1,2,3,4,5))
# Default transformation can improve performance.
N = 100
K = 1000
X = matrix(exp(rnorm(N*K)), nrow=N, ncol=K)
system.time(coordinates(X, alr_basis(K)))
system.time(coordinates(X, 'alr'))
}
\seealso{
See functions \code{\link{ilr_basis}}, \code{\link{alr_basis}},
\code{\link{clr_basis}}, \code{\link{sbp_basis}}
to define different compositional basis.
See function \code{\link{composition}} to obtain details on how to calculate
a compositions from given coordinates.
}
