% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist}
\alias{dist}
\title{Distance Matrix Computation (including Aitchison distance)}
\usage{
dist(x, method = "euclidean", ...)
}
\arguments{
\item{x}{A data matrix whose rows are compositions.}

\item{method}{The distance measure to be used. This must be one of
\code{"aitchison"}, \code{"euclidean"}, \code{"maximum"},
\code{"manhattan"}, \code{"canberra"}, \code{"binary"}, or
\code{"minkowski"}. Any unambiguous abbreviation can be given.}

\item{...}{Additional arguments passed to \code{\link[stats]{dist}}.}
}
\value{
An object of class \code{"dist"}.
}
\description{
Compute a distance matrix for compositional data, including the Aitchison
distance as an extension of \code{\link[stats]{dist}}.
}
\examples{
X <- exp(matrix(rnorm(10 * 50), ncol = 50, nrow = 10))

(d <- dist_coda(X, method = "aitchison"))
plot(hclust(d))

# In contrast to Euclidean distance
dist(rbind(c(1, 1, 1), c(100, 100, 100)), method = "euc")

# Using Aitchison distance, only relative information is of importance
dist_coda(rbind(c(1, 1, 1), c(100, 100, 100)), method = "ait")

}
\seealso{
\code{\link{dist_coda}}, \code{\link[stats]{dist}}
}
