% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eurostat_employment}
\alias{eurostat_employment}
\title{Employment distribution in EUROSTAT countries}
\format{
An object of class \code{data.frame} with 29 rows and 17 columns.
}
\usage{
eurostat_employment
}
\description{
Employment distribution in EUROSTAT countries
}
\keyword{datasets}
