% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foraminiferals}
\alias{foraminiferals}
\title{Paleocological compositions}
\format{
An object of class \code{data.frame} with 30 rows and 5 columns.
}
\usage{
foraminiferals
}
\description{
Paleocological compositions
}
\keyword{datasets}
