% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_expend}
\alias{house_expend}
\title{Household expenditures}
\format{
An object of class \code{data.frame} with 27 rows and 15 columns.
}
\usage{
house_expend
}
\description{
Household expenditures
}
\keyword{datasets}
