% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{household_budget}
\alias{household_budget}
\title{Household budget patterns}
\format{
An object of class \code{data.frame} with 40 rows and 6 columns.
}
\usage{
household_budget
}
\description{
In a sample survey of single persons living alone in rented accommodation, twenty
men and twenty women were randomly selected and asked to record over a period of
one month their expenditures on the following four mutually exclusive and
exhaustive commodity groups:
* Hous: Housing, including fuel and light.
* Food: Foodstuffs, including alcohol and tobacco.
* Serv: Services, including transport and vehicles.
* Other: Other goods, including clothing, footwear and durable goods.
}
\keyword{datasets}
