% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{ilr_basis}
\alias{ilr_basis}
\title{Build an isometric log-ratio basis}
\usage{
ilr_basis(dim)
}
\arguments{
\item{dim}{number of components}
}
\value{
matrix
}
\description{
Build an isometric log-ratio basis
}
\examples{
ilr_basis(5)
}
