% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals_milk}
\alias{mammals_milk}
\title{Mammal’s milk}
\format{
An object of class \code{data.frame} with 24 rows and 6 columns.
}
\usage{
mammals_milk
}
\description{
Mammal’s milk
}
\keyword{datasets}
