% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals_milk}
\alias{mammals_milk}
\title{Mammal’s milk}
\format{
An object of class \code{data.frame} with 24 rows and 6 columns.
}
\usage{
mammals_milk
}
\description{
The mammalsmilk data set contains the percentages of five constituents (W: water,
P: protein, F: fat, L: lactose, and A: ash) of the milk of 24 mammals. The data are
taken from [Har75].
}
\keyword{datasets}
