% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk_cows}
\alias{milk_cows}
\title{Milk composition study}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 116 rows and 10 columns.
}
\usage{
milk_cows
}
\description{
In an attempt to improve the quality of cow milk, milk from each of thirty cows
was assessed by dietary composition before and after a strictly controlled dietary
and hormonal regime over a period of eight weeks. Although seasonal variations in
milk quality might have been regarded as negligible over this period, it was decided
to have a control group of thirty cows kept under the same conditions but on a
regular established regime. The sixty cows were of course allocated to control and
treatment groups at random. The `milk_cows` data set provides the complete set of
before and after milk compositions for the sixty cows, showing the protein (pr),
milk fat (mf), carbohydrate (ch), calcium (Ca), sodium (Na) and potassium (K)
proportions by weight of total dietary content.
}
\keyword{datasets}
