% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{milk_cows}
\alias{milk_cows}
\title{Milk composition study}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 116 rows and 10 columns.
}
\usage{
milk_cows
}
\description{
Milk composition study
}
\keyword{datasets}
