% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{montana}
\alias{montana}
\title{Concentration of minor elements in carbon ashes}
\format{
An object of class \code{data.frame} with 229 rows and 6 columns.
}
\usage{
montana
}
\description{
Concentration of minor elements in carbon ashes
}
\keyword{datasets}
