% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{pairwise_basis}
\alias{pairwise_basis}
\title{Pairwise log-ratio generator system}
\usage{
pairwise_basis(dim)
}
\arguments{
\item{dim}{An integer indicating the number of components.
If a dataframe or matrix is provided, the number of components is inferred from the number of columns. If a character vector specifying the names of the parts is provided the number of component is its length.}
}
\value{
matrix
}
\description{
The function returns all combinations of pairs of log-ratios.
}
