% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{pairwise_basis}
\alias{pairwise_basis}
\title{Pairwise log-ratio generator system}
\usage{
pairwise_basis(dim)
}
\arguments{
\item{dim}{dimension to build the pairwise log-ratio generator system}
}
\value{
matrix
}
\description{
The function returns all combinations of pairs of log-ratios.
}
