% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parliament2017}
\alias{parliament2017}
\title{Results of catalan parliament elections in 2017 by regions.}
\format{
A data frame with 42 rows and 9 variables:
\describe{
\item{com}{Region}
\item{cs}{Votes to Ciutadans party}
\item{jxcat}{Votes to Junts per Catalunya party}
\item{erc}{Votes to Esquerra republicana de Catalunya party}
\item{psc}{Votes to Partit socialista de Catalunya party}
\item{catsp}{Votes to Catalunya si que es pot party}
\item{cup}{Votes to Candidatura d'unitat popular party}
\item{pp}{Votes to Partit popular party}
\item{other}{Votes to other parties}
}
}
\source{
\url{https://www.idescat.cat/tema/elecc}
}
\usage{
parliament2017
}
\description{
Results of catalan parliament elections in 2017 by regions.
}
\keyword{datasets}
