% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{pc_basis}
\alias{pc_basis}
\title{Isometric log-ratio basis based on Principal Components.}
\usage{
pc_basis(X)
}
\arguments{
\item{X}{compositional dataset}
}
\value{
matrix
}
\description{
Different approximations to approximate the principal balances of a compositional dataset.
}
