% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{petrafm}
\alias{petrafm}
\title{Calc-alkaline and tholeiitic volcanic rocks}
\format{
An object of class \code{data.frame} with 100 rows and 4 columns.
}
\usage{
petrafm
}
\description{
\loadmathjax

This petrafm data set is formed by 100 classified volcanic rock samples from
Ontario (Canada). The three parts are:

\mjeqn{[A: {Na}_2 O + K_2 O; F: Fe O + 0.8998 Fe_2 O_3 ; M: Mg O]}{[A: Na2 O + K2 O; F: FeO + 0.8998 · Fe2 O3 ; M: MgO]}

Rocks from the calc-alkaline magma series (25) can be well distinguished from
samples from the tholeiitic magma series (75) on an AFM diagram.
}
\keyword{datasets}
