% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{petrafm}
\alias{petrafm}
\title{Calc-alkaline and tholeiitic volcanic rocks}
\format{
An object of class \code{data.frame} with 100 rows and 4 columns.
}
\usage{
petrafm
}
\description{
Calc-alkaline and tholeiitic volcanic rocks
}
\keyword{datasets}
