% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{plot_balance}
\alias{plot_balance}
\title{Plot a balance}
\usage{
plot_balance(B, data = NULL, main = "Balance dendrogram", ...)
}
\arguments{
\item{B}{Balance to plot}

\item{data}{(Optional) Data used to calculate the statistics associated to a balance}

\item{main}{Plot title}

\item{...}{further arguments passed to plot}
}
\value{
Balance plot
}
\description{
Plot a balance
}
