% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pottery}
\alias{pottery}
\title{Chemical compositions of Romano-British pottery}
\format{
An object of class \code{data.frame} with 45 rows and 11 columns.
}
\usage{
pottery
}
\description{
The pottery data set consists of data pertaining to the chemical composition of 45
specimens of Romano-British pottery. The method used to generate these data is
atomic absorption spectophotometry, and readings for nine oxides
(Al2O3, Fe2O3, MgO, CaO, Na2O, K2O, TiO2 , MnO, BaO) are provided. These samples come
from five different kiln sites.
}
\keyword{datasets}
