% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{serprot}
\alias{serprot}
\title{Serum proteins}
\format{
An object of class \code{data.frame} with 36 rows and 7 columns.
}
\usage{
serprot
}
\description{
The `serprot` data set records the percentages of the four serum proteins
from the blood samples of 30 patients. Fourteen patients have one
disease (1) and sixteen are known to have another different disease (2).
The 4-compositions are formed by the proteins [albumin, pre-albumin,
globulin A, globulin B].
}
\keyword{datasets}
