% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statistitian_time}
\alias{statistitian_time}
\title{A statistician’s time budget}
\format{
An object of class \code{data.frame} with 20 rows and 7 columns.
}
\usage{
statistitian_time
}
\description{
A statistician’s time budget
}
\keyword{datasets}
