% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variation_array}
\alias{variation_array}
\title{Variation array is returned.}
\usage{
variation_array(X, only_variation = FALSE)
}
\arguments{
\item{X}{Compositional dataset}

\item{only_variation}{if TRUE only the variation matrix is calculated}
}
\value{
variation array matrix
}
\description{
Variation array is returned.
}
\examples{
set.seed(1)
X = matrix(exp(rnorm(5*100)), nrow=100, ncol=5)
variation_array(X)
variation_array(X, only_variation = TRUE)
}
