% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waste}
\alias{waste}
\title{The waste composition in Catalonia}
\format{
An object of class \code{data.frame} with 215 rows and 10 columns.
}
\usage{
waste
}
\description{
The waste composition in Catalonia
}
\keyword{datasets}
