% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weibo_hotels}
\alias{weibo_hotels}
\title{Hotel posts in social media}
\format{
An object of class \code{data.frame} with 10 rows and 5 columns.
}
\usage{
weibo_hotels
}
\description{
Hotel posts in social media
}
\keyword{datasets}
